requirements.txt
Django
requests
python-dotenv
fuzzywuzzy
asgiref
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary